--- app/Http/Controllers/ExpenseTemplateController.php
@@ -1,85 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\ExpenseTemplate;
-use Illuminate\Http\Request;
-
-class ExpenseTemplateController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\ExpenseTemplate  $expenseTemplate
-     * @return \Illuminate\Http\Response
-     */
-    public function show(ExpenseTemplate $expenseTemplate)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\ExpenseTemplate  $expenseTemplate
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(ExpenseTemplate $expenseTemplate)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\ExpenseTemplate  $expenseTemplate
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, ExpenseTemplate $expenseTemplate)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\ExpenseTemplate  $expenseTemplate
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(ExpenseTemplate $expenseTemplate)
-    {
-        //
-    }
-}
